Wrote a trait for filters which can filter on a node name.
